Arithmetic expression evaluation is now working.
Fifth version of the GUI with all panels in order. It is taken the filename from the first dialog input box and displays the file and gives the result accordingly.

Trying to write the output to a file so it has been modified the lisp.g but it throws some errors when building.